pyproject.toml
[project]
name = "mcptool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.22",
    "mcp-use>=1.3.0",
    "requests",
    "python-dotenv",
    "langchain-groq>=0.3.2",
    "langchain-core",
]
